backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (PostgreSQL)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Request handling
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
